Add modify support for private, status, assignee, flags, and cc
These are options that I needed for my workflow so I added them to the
bugzilla tool.

private - just sets a flag to distinguish private/non-private comments.
this is passed through to the backend api logic

status - changes the status of a bug.  utilizes the setstatus api

assignee - changes the assignee of a bug.  utilizes the setassignee api

flag - sets the flags in the bugzilla and utilizes the new updateflags api.
multiple flags will need to use multiple '-f's, no comma support.  the code
then cycles through the list of -f's and creates a list to pass down

cc - sets the cc in the bugzilla to the provide email address using the
addcc api.  same as the flags above, multiple cc's need multiple --cc's.

All have been tested with my test bugzilla and some with my normal workflow