Add support for adding comments to bugzilla modifications
"Most modifications commands allow you to add comments to your bugzilla
changes.  However at the same time, one can add a stand-alone comment to the
bugzilla.  The question is how to do this without creating a mess for an
api.

The changes below piggy-back the --comment option, but saying that if no
other command is using the comment then it has to be a stand-alone one.
Otherwise pass the comment into the command (if it supports it).

In this particular change, I added support for the 'close' command."

From Don Zickus - thanks!