bin/bugzilla: add --fixed_in option to 'modify' command
This uses the _update_bugs() method to implement it. I know it is not the
recommended way to do that, but that's the only way provided by the API to set
the "fixed in" field without closing the bug (on my case, I always set the
field when changing the bug state to MODIFIED, not when marking it as CLOSED).

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>