Implement unicode(bug)
The bug summary is an unicode string, not a byte string (a str object), and
__str__() was returning an unicode string for all bugs. This patch moves the
__str__() code to __unicode__() and just use unicode(bug) on __str__() to keep
the current behavior.

We can't just return unicode objects on __str__() because calling
"unicode(bug)" would contain a implicit conversion to str[1], that would break
depending on the system default encoding.

Also, "print bug" will convert the return value of __str__() to str before
printing it, potentially causing encoding conversion errors. That will be
addressed on a further commit.

[1] It is equivalent to calling unicode(str(bug.__str__()))

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>