Add more query functionality to bugzilla flags
With the current abuse of bugzilla flags, one really needs the ability to
properly query them.  These changes allow multiple '--flag' options and
gives the user the ability to use '&,|' options with the flags.

For example, now you can do
--flag 'rhel-5.4.0? | rhel-5.4.0+' --flag 'blocker? | blocker+'

Now the ability to separate the flag commands like above is important for
the negation command later, where negation only works on the flag group it
is apart of.

Tested with RHEL-5 queries.