Add negate feature to bugzilla query
Sometimes it is nice to negate a certain boolean query for the results you
are looking for to help filter stuff you don't care about.  This adds a new
symbol '!' to be used for that purpose.

This only works in the boolean query group for which the negation is apart
of.  Subsequent boolean queries will not use the negation.

For example:
Show me everything in the kernel that is MODIFIED by does not have the
rhel-5.3.z flag and has the blocker flag set

bugzilla query -p '..RHEL-5..' -c kernel -t MODIFIED --flag '! rhel-5.3.z? |
rhel-5.3.z+' --flag 'blocker+'