Handle really simple boolean queries
There are some boolean queries that do not have values associated with them,
but the bugzilla scripts expects it.  Modify the code to check to see if a
third argument exists, if so use the value field, otherwise ignore it.

This handles the case of 'cf_release_notes-isnull'.

Thanks to Jeff Moyer for this fix.