NovellBugzilla implementation.
The NovellBugzilla implementation - is a subclass of Bugzilla32 with
reimplemented _login and _logoout methods compatible with iChain.

NovellBugzilla don't allow other self.url than bnc, because it should
not be used for any other bugzilla. The url parameters in __init__ and
connect() are overwritten and exists only for compatibility purposes.

It can also read the username/password from ~/.oscrc, which is
common for many SUSE users and contains a same iChain credentials as is
necessary for bugzilla login. So when user use osc he don't need
duplicate login informations to ~/.bugzillarc.