Change the --oneline flag output to remove the ':'
Don Zickus said:
I noticed the bugzilla emails use special email headers that include what the
flag states are. In order to be consistent, I decided to remove the ':' part of
the output and instead have the flag and value go together.

For example: devel_ack:? --> devel_ack?

I don't think too many people are using that option yet so it shouldn't cause
any confusion.

This changes the output of the CLI, possibly breaking scripts, so it probably
shouldn't get applied to the 0.5 bugfix branch.