bugzilla4: use query_format = advanced
The old bugzilla was loose in the advanced query requirement. The
new bugzilla expects the q['query_format'] = advanced now.  The old
way should be able to handle that too.  Hence this change should be
backwards compatible.

I could do this in the backend too like some of the other changes,
but the frontend should have really been passing the argument.  It
can go either way.

Signed-off-by: Don Zickus <dzickus@redhat.com>