bugzilla4: speed up the normal query case
The normal bugzilla query case gets only the bz id back, causing
a cache miss in the backend to fetch the bugzilla supported fields.  In
the old bugzilla, this fetch took a couple of seconds.  With bugzilla4, this
fetch is 50 seconds.  To workaround this new limitation, explictly fetch the
fields needed in the normal output case.

Signed-off-by: Don Zickus <dzickus@redhat.com>