Add a simple query test
Do some module symlink hackery that allows us to import bin/bugzilla,
and run main() directly, so we can get python objects returned. Hack
bin/bugzilla a bit to give better unittest behavior if a unittest env
variable is set.

This is a huge hack, but my motivation is to first test existing
functionality before I start refactoring, and this is the easiest way
that I can think of.