bugzilla: use passed in cookies correctly
When using the option --cookiefile for bugzilla, the cookiefile
override happens to late.  Use the passed in file during the
beginning of init.

This is addressed by defaulting the cookiefile to None to simplify
the passed in argument code.  Then the Bugzilla init routine assigns
the default of ~/.bugzillacookies if cookiefile is None.

This shouldn't affect any of the other scripts that do not use the
cookiefile as an arg.

Reported-by: Jan Pokorný <jpokorny@redhat.com>
Signed-off-by: Don Zickus <dzickus@redhat.com>