Add 'bugzilla new --groups'
Since BZ 4.0 there is an XMLRPC/Web service capability to define
groups when creating a bug.  There are additional capabilities to add
or remove groups, but to keep this simple for now, I just stuck with
the create time function.

I verified that *not* setting a group seems to do nothing different
than in the current released version (0.8.0) of the client.  Patch
attached against git master.