Remove NovellBugzilla
A Novell person hasn't touched this code in 4 years. I can't get it to
work. And future changes are about to break things even more.

I'm happy to help anyone revive this code if need be, but for now
I'm dropping it. git will make it trivial to revive if needed.