CVE-2013-2191: Switch to pycurl to get SSL host and cert validation
Right now python-bugzilla will happily allow connecting to a host with
a self signed SSL certificate, or hostname that doesn't match the
cert. This isn't a safe default.

Standard python libs don't handle these cases, but pycurl does. So
we switch to pycurl for the transport layer.

Add a --nosslverify CLI switch to turn off this functionality if the
user chooses.

Thanks to Tomas Hoger for much of the sample code.