bugzilla: Switch bug_autorefresh default to False
The bug_autorefresh behavior is not a good idea as a default. It's
been this way forever but it's time to change. While it makes writing
the code slightly easier, it's really easy to make scripts incredibly
non-performant scripts, which has come up in many different instances.

Extend the error message for missing attributes to try and hint at
the solution as well.

https://bugzilla.redhat.com/show_bug.cgi?id=1297637