Client side certificate support
Adds the `cert` command line and configuration options that allow you to
specify a client-side certificate file to use for TLS authentication
against the web server.

File format should be concatenated certificate and (unencrypted) private
key. Encrypted private key should also be possible, but would have to
deal with passing the password along to `requests`, and I'm not
comfortable enough with Python to actually facilitate that.