Add `login --api-key` support for bugzilla login API
This change introduces a new feature for the command line client
and the Bugzilla python API.

In the python API, it changes the interactive_login method
to receive a new flag: "use_api_key" that, in case is defined will
ask the user to provide an API key instead of regular username/password.

In case the API_KEY provide works correctly, it will also
update the ~/config/python-bugzilla/bugzillarc file (or ~/.bugzillarc).

In the bugzilla-cli side, it will add a new parameter to the login
command: "--api-key". Use key will trigger the usage of this new python API.

Default behaviors were not change, neither in bugzilla-cli or the python
API.

Resolves: #82