cli: Add query --extrafield, --includefield, --excludefield
Allow users to pass these extra/include/exclude values straight
through to the remote server.

For most of the output format options these don't add anything,
but for --json and --raw which do a getbug() call, these can be useful
for tweaking the output, like reducing the total returned data, or
requesting the server return a non-standard field.

Signed-off-by: Cole Robinson <crobinso@redhat.com>