session: Rework how auth params are passed to backends
REST backend relies on session to set the params, but XMLRPC does
it in the backend. Switch it so session hands a param dictionary off
to the backend.

This will help with upcoming changes that make param values more
conditional, and a possible future where bugzilla doesn't accept
auth via URL params

Signed-off-by: Cole Robinson <crobinso@redhat.com>