base: Explicitly check for None instead of falsy value
In `_getbugs()`, the bit of code that ensures that the output order of
the bugs is the same as the input order of ids incorrectly checks
whether idint / aliaslist are falsy instead of explicitly checking for
None.

If idlist / aliaslist contain falsy non-None values such as 0 or "",
they will skip the check and be potentially incorrectly added to the
return list.

An example of where this fails would be an idlist = range(0, 20)
From tests against the Red Hat bugzilla instance, 0 is not a valid
bug_id thus the API will return 1-20, but since idval = 0 will skip the
two checks, the end result will be of length 21 (unexpected) and the
first bug will be in the return list twice.