Provide 'total_matches' in query response if server provides it
The RH Bugzilla provides a 'total_matches' value in a query response
which the querier can use to know whether the response included all
matching bugs or not. This patch adds a proxy object wrapped around
the list of returned bugs to expose this value if the server has
provided it. The proxy object implements enough of the Python
iterator protocol so that it can be used as a replacement for existing
code issuing queries, as long as that code does not attempt to modify
the returned list object.