build_update: don't convert 'blocks' or 'depends' to int
This is not necessary - Bugzilla, at least the current version
on bugzilla.redhat.com, is happy to accept IDs as a string. It
also causes a problem: this prevents you from using aliases when
setting blocks or depends, which is often very convenient when
bugs have predictable aliases (like "F40Changes" or
"F39BetaBlocker"). Without this change, you have to do an extra
query just to find the ID of the bug.

Signed-off-by: Adam Williamson <awilliam@redhat.com>