examples: add redhat_query_all.py
In late 2021, bugzilla.redhat.com changed query() results to default
returning only 20 bugs. If the user passes in limit=0, that number
changes to 1000, but is still capped if the query would return more
than that.

There's a discussion here with multiple proposed workarounds:
https://github.com/python-bugzilla/python-bugzilla/issues/149

This demonstrates the one that takes the least amount of code IMO.
It uses `ids_only=True`, which is a custom bugzilla.redhat.com
query feature to bypass the query limit by only returning matching
bug IDs.

Signed-off-by: Cole Robinson <crobinso@redhat.com>