base: Add query_return_extra
This is like `query()`, but the return value is altered to be
(buglist, values), where `values` is raw dictionary output from
the API call, excluding the bug content. For example this may
include a `limit` value if the bugzilla instance puts an implied
limit on returned result numbers.

bugzilla.redhat.com also has a custom extension to report
`total_matches` for a query, which lets user know if the returned
results are complete or not

Signed-off-by: Cole Robinson <crobinso@redhat.com>