fix: Emulate `permissive` on REST backend (closes #222)
`_BackendREST.bug_get` cannot pass a `permissive` parameter
to the server as the REST API does not honor such a
parameter.

With this change, permissiveness is handled inside the
method:

If `permissive` is false and a single ID or alias is requested, the "get" method is used and an exception gets
raised in case of error.
Otherwise, the "search" method is used, which may return
an empty list, if the client is not authenticated or an
ID or alias does not exist.