Rename method getcomments to get_comments for better readablity (#226)
Using the api, I used to think the only way to retrieve the comments is
to go through a Bugzilla instance. Later, I figured out the the Bug
objects have the method `get_attachments` and I was amazed that there is
no `get_comments`. Digging into the code I realized the method has been
named `getcomments`. I have changed all the occurrences of `getcomments`
method of the Bug class to `get_comment`

---------

Co-authored-by: abahrani <abahrani@starfishmedical.com>
Co-authored-by: Andreas Hasenkopf <andreas@hasenkopf.xyz>